Shrinker fixes: adoption semantics, live-state walks, dead searches
- consider() now returns whether the candidate became (or already was)
  the shrink target, matching Hypothesis's consider_new_nodes. Passes
  that keep a local view of the target in sync with the return value
  (reorder_spans/shrink_ordering, try_sort_group, remove_discarded's
  loop condition) no longer desync on interesting-but-not-adopted runs,
  and remove_discarded's termination follows directly.
- redistribute_string_pairs / redistribute_bytes_pairs extend an
  accepted single-element move with find_integer against live state.
  The old bin_search_down_r re-probed the already-current state, so its
  first candidate short-circuited true and the multi-element search was
  dead — one element per pass invocation instead of O(log k).
- node_program's leftward walk probes the live shrink target, so
  accepted steps compound (Hypothesis's offset_left); against the old
  fixed snapshot the second step re-included the deleted region and the
  walk stalled after one step.
- shrink_strings truncates with the same binary search + bounded scan
  shrink_bytes uses, instead of one execution per candidate length.
- probe() only treats a stall as terminal after at least one
  improvement, matching consider(), so the mutation/coarse probes are
  no longer permanently disabled in zero-improvement shrinks.
- lower_and_bump issues its gap-independent candidates once per node
  rather than once per gap; shrink_duplicates' doubling step no longer
  overflows u64 in debug builds; the direct-execution sites in
  try_trivial_spans and NaN canonicalization count toward calls.

+++ hegel-c/src/native/shrinker/integers.rs
@@ -500,7 +500,7 @@ impl<'a> Shrinker<'a> {
                 }
                 if live_base(self) > lo {
                     find_integer_r(|n| {
-                        let attempt = live_base(self) - BigInt::from(2u64 * n as u64);
+                        let attempt = live_base(self) - BigInt::from(2u128 * n as u128);
                         group_replace(self, &attempt)
                     })?;
                 }
@@ -523,7 +523,7 @@ impl<'a> Shrinker<'a> {
                 let neg_hi = -&lo;
                 if live_base(self) < neg_hi {
                     find_integer_r(|n| {
-                        let attempt = live_base(self) + BigInt::from(2u64 * n as u64);
+                        let attempt = live_base(self) + BigInt::from(2u128 * n as u128);
                         group_replace(self, &attempt)
                     })?;
                 }
